Change Hunk_* and zone alloc functions to use size_t instead of int. Replace READ_COMPRESSED_SAVEGAMES and WRITE_COMPRESSED_SAVEGAMES macros with COMPRESSED_SAVEGAMES. Misc changes and some cleanup to game code.

+++ game/g_turret.c
@@ -173,7 +173,7 @@ void hrocket_turret_fire (edict_t *self, edict_t *owner, vec3_t start, vec3_t di
 }
 //CW--
 
-static unsigned long HB_Shots;
+static unsigned int /*long*/ HB_Shots;
 
 void turret_breach_fire (edict_t *self)
 {
